+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Changelog
 
-### Version 0.27.6-alpha:
+### Version 0.28.7-alpha:
 - Added a function to install and convert server packs.
 - Added a function to replace files.
 - Added icons to every Button on the Toolbar of the text-editor.
@@ -19,6 +19,8 @@
 - Replaced the about window with a new one.
 - Changed the mod detector to show mod names instead of file names.
 - Added a function to install zip packs.
+- Added tiling to the image-viewer.
+- Changed tiling to two separate options.
 
 ### Version 0.17.6-alpha:
 - Piped `.json` files through `jsonesque` first, so comments can exist in the `.json` files.